templates/cybereng/v1/requirements.txt
altgraph==0.17.4
anyio==4.7.0
asttokens==3.0.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
decorator==5.1.1
docutils==0.21.2
executing==2.1.0
github3.py==4.0.1
h11==0.14.0
idna==3.10
iniconfig==2.0.0
ipython==8.31.0
itsdangerous==2.2.0
jedi==0.19.2
marimo==0.10.9
Markdown==3.7
markdown-it-py==3.0.0
matplotlib-inline==0.1.7
mdurl==0.1.2
narwhals==1.20.1
packaging==24.2
parse_pip_search==0.0.1
parso==0.8.4
pefile==2024.8.26
pluggy==1.5.0
prompt_toolkit==3.0.48
psutil==6.1.1
pure_eval==0.2.3
pycparser==2.22
Pygments==2.18.0
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.11
PyJWT==2.10.1
pymdown-extensions==10.13
pytest==8.3.4
python-dateutil==2.9.0.post0
pywin32-ctypes==0.2.3
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
ruff==0.8.4
setuptools==75.1.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
starlette==0.45.1
tomlkit==0.13.2
traitlets==5.14.3
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
websockets==14.1
wheel==0.44.0
